% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GO.BP.list_mus}
\alias{GO.BP.list_mus}
\title{GO_BP geneset for mouse}
\format{
An object of class \code{list} of length 7526.
}
\usage{
GO.BP.list_mus
}
\description{
GO_BP geneset for mouse
}
\details{
GO_BP terms from msigdb
}
\keyword{datasets}
