% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MmCMS.R
\name{MmCMS}
\alias{MmCMS}
\title{CMS classification}
\usage{
MmCMS(
  emat,
  templates = MmCMS::template.CMS.A,
  Genesets = c("template.CMS.B", "template.CMS.C", "template.CMS.A"),
  nPerm = 1000,
  seed = NULL,
  FDR = 0.05,
  doPlot = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{emat}{a numeric expression matrix with sample columns, and HGNC symbol rownames.
Data should be normalized.
see the example in \code{\link{TestData_gemm}}.}

\item{templates}{a data frame with two columns; \emph{class} (coerced to
factor) and \emph{probe} (coerced to character). There are 3 templates: 'template.CMS.A', 'template.CMS.B' and 'template.CMS.C'.
Default is template.CMS.A}

\item{Genesets}{Select the same as template}

\item{nPerm}{an integer, number of permutations for \eqn{p}-value
estimation.}

\item{seed}{an integer, for \eqn{p}-value reproducibility.}

\item{FDR}{a false discovery rate, sets prediction confidence threshold.}

\item{doPlot}{a logical, whether to produce prediction subHeatmap.}

\item{verbose}{a logical, whether console messages are to be displayed.}
}
\value{
a data frame with class predictions, template distances,
\eqn{p}-values and false discovery rate adjusted \eqn{p}-values
(\code{\link[stats]{p.adjust}}). Rownames equal emat
colnames.
}
\description{
consensus molecular subtypes (CMS) classification for mouse tissues.
}
\details{
\code{MmCMS} provides 3 options to call CMS subtypes in mouse tissues.
The core algorithm is the \href{https://github.com/peterawe/CMScaller}{CMScaller} developed by Eide PW, et al. (2017)
and the Nearest Template Prediction (NTP) algorithm as proposed by Yujin Hoshida (2010).
}
\note{
genes with missing values are discarded.
}
\examples{
emat <- TestData_gemm
re <- MmCMS(emat, templates=MmCMS::template.CMS.A, Genesets = c("template.CMS.A"), seed=120)
}
\references{
Eide PW, Bruun J, Lothe RA, Sveen A. (2017). CMScaller:
an R package for consensus molecular subtyping of colorectal cancer pre-clinical models.
doi: 10.1038/s41598-017-16747-x.

Hoshida, Y. (2010). Nearest Template Prediction: A
Single-Sample-Based Flexible Class Prediction with Confidence Assessment.
PLoS ONE 5, e15543.

Guinney J, Dienstmann R, Wang X, de Reynies A, Schlicker A,
Soneson C, et al. The consensus molecular subtypes of colorectal cancer.
Nat Med. 2015;21:1350-6.
}
\seealso{
\code{\link{template.CMS.A}}, \code{\link{template.CMS.B}}, \code{\link{template.CMS.C}}
}
