% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{option.C.geneset.list_mus}
\alias{option.C.geneset.list_mus}
\title{option.C geneset for mouse}
\format{
An object of class \code{list} of length 77.
}
\usage{
option.C.geneset.list_mus
}
\description{
option.C geneset for mouse
}
\details{
gene set includes a combined pathway system
}
\keyword{datasets}
