% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{template.CMS.A}
\alias{template.CMS.A}
\title{Mouse CMS template (option A)}
\format{
An object of class \code{data.frame} with 533 rows and 2 columns.
}
\usage{
template.CMS.A
}
\description{
Mouse CMS template (option A)
}
\details{
CMS prediction template for NTP method. Marker genes were
identified by converting human CMS template from \href{https://www.nature.com/articles/s41598-017-16747-x}{CMScaller} to mouse orthologe using biomaRt.
\code{templates$probe} refers to gene symbols.
}
\examples{
emat <- TestData_gemm
re <- MmCMS(emat, templates=MmCMS::template.CMS.A, Genesets = c("template.CMS.A"), seed=120)
}
\references{
Eide PW, Bruun J, Lothe RA, Sveen A. CMScaller:
an R package for consensus molecular subtyping of colorectal cancer pre-clinical models. Scientific reports (2017)

Guinney J, Dienstmann R, Wang X, de Reynies A, Schlicker A, Soneson C, et al. The consensus molecular subtypes of colorectal cancer. Nat Med (2015)

Hoshida Y. Nearest Template Prediction: A Single-Sample-Based Flexible Class Prediction with Confidence Assessment. PLoS ONE (2010)

Durinck S, Spellman P, Birney E, Huber W (2009). “Mapping identifiers for the integration of genomic datasets with the R/Bioconductor package biomaRt.” Nature Protocols, 4, 1184–1191.
}
\keyword{datasets}
