% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{template.CMS.B}
\alias{template.CMS.B}
\title{Mouse CMS template (option B)}
\format{
An object of class \code{data.frame} with 974 rows and 2 columns.
}
\usage{
template.CMS.B
}
\description{
Mouse CMS template (option B)
}
\details{
CMS prediction template for NTP method. Marker genes were
identified via a biological knowledge-based approach instead of having sole emphesis on individual genes.
ssGSEA was performed on TCGA RNA-sequencing data from original \href{https://pubmed.ncbi.nlm.nih.gov/26457759/}{CMSclassifier}
with assigned CMS subtypes, to provide GO_BP scores for each sample.
Then CMS-related GO_BP terms were converted to mouse GO_BP terms orthologe using msigdbr.
Instead of individual genes, the GO_BP terms used as a template.
\code{templates$probe} refers to GO_BP terms.
}
\examples{
emat <- TestData_gemm
re <- MmCMS(emat, templates=MmCMS::template.CMS.B, Genesets = c("template.CMS.B"), seed=120)
}
\references{
Eide PW, Bruun J, Lothe RA, Sveen A. CMScaller:
an R package for consensus molecular subtyping of colorectal cancer pre-clinical models. Scientific reports (2017)

Guinney J, Dienstmann R, Wang X, de Reynies A, et al. The consensus molecular subtypes of colorectal cancer. Nat Med (2015)

Hoshida Y. Nearest Template Prediction: A Single-Sample-Based Flexible Class Prediction with Confidence Assessment. PLoS ONE. 2010;5:e15543.

Liberzon et al. (2015) The Molecular Signatures Database Hallmark Gene Set Collection. DOI:https://doi.org/10.1016/j.cels.2015.12.004

Hänzelmann S., Castelo R. and Guinney J. GSVA: gene set variation analysis for microarray and RNA-Seq data. BMC Bioinformatics, 14:7, 2013.
}
\keyword{datasets}
