% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TestData_gemm}
\alias{TestData_gemm}
\title{Expression profile of GEMMs}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 24751 rows and 18 columns.
}
\usage{
TestData_gemm
}
\description{
Expression profile of GEMMs
}
\details{
18 GEMM data from Cancer Cell. 2019 Sep 16;36(3):319-336.e7.doi: 10.1016/j.ccell.2019.08.003.
}
\keyword{datasets}
